+++ outputs.tf
@@ -17,3 +17,8 @@ output "location_slug" {
   value       = local.location_slug
   description = "Azure region in slug format"
 }
+
+output "location_paired" {
+  description = "Azure paired region with the current one (in slug format)"
+  value       = local.paired[local.location_slug]
+}